Move holdings util to holdings.go
