+++ broker/catalog/creator_impl.go
@@ -2,11 +2,9 @@ package catalog
 
 import (
 	"fmt"
-	"strings"
 
 	"github.com/indexdata/crosslink/broker/ill_db"
 	"github.com/indexdata/crosslink/directory"
-	"github.com/indexdata/crosslink/iso18626"
 )
 
 const (
@@ -94,25 +92,4 @@ func (c *LookupAdapterCreatorImpl) GetAdapter(peer ill_db.Peer) (LookupAdapter, 
 		}
 	}
 	return nil, fmt.Errorf("must specify either sru or zoom properties for lookup adapter type")
-}
-
-func LookupParamsFromBibliographicInfo(info iso18626.BibliographicInfo, serviceInfo *iso18626.ServiceInfo) LookupParams {
-	var serviceType string
-	if serviceInfo != nil {
-		serviceType = string(serviceInfo.ServiceType)
-	}
-	params := LookupParams{
-		Identifier:  info.SupplierUniqueRecordId,
-		Title:       info.Title,
-		ServiceType: serviceType,
-	}
-	for _, id := range info.BibliographicItemId {
-		switch strings.TrimSpace(id.BibliographicItemIdentifierCode.Text) {
-		case "ISBN":
-			params.Isbn = id.BibliographicItemIdentifier
-		case "ISSN":
-			params.Issn = id.BibliographicItemIdentifier
-		}
-	}
-	return params
 }
